# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DiTing Studio API",
    "description": "DiTing Studio API — chat sessions, scheduled jobs, platform channels, model management, skills, memory, logs, file browser, group chat, and terminal.",
    "version": "0.6.20"
  },
  "servers": [
    {
      "url": "http://localhost:8648",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/DiTing/group-chat/rooms": {
      "post": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Create rooms",
        "description": "POST /api/DiTing/group-chat/rooms",
        "operationId": "createRooms",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agents": {
                    "type": "boolean"
                  },
                  "compression": {
                    "type": "number"
                  },
                  "inviteCode": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "inviteCode",
                  "name"
                ]
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Get rooms",
        "description": "GET /api/DiTing/group-chat/rooms",
        "operationId": "getRooms",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/DiTing/group-chat/rooms/join/{code}": {
      "get": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Get :code",
        "description": "GET /api/DiTing/group-chat/rooms/join/:code",
        "operationId": "getJoin",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/DiTing/group-chat/rooms/{roomId}": {
      "get": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Get :roomId",
        "description": "GET /api/DiTing/group-chat/rooms/:roomId",
        "operationId": "getRooms",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Delete :roomId",
        "description": "DELETE /api/DiTing/group-chat/rooms/:roomId",
        "operationId": "deleteRooms",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/DiTing/group-chat/rooms/{roomId}/agents": {
      "post": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Create agents",
        "description": "POST /api/DiTing/group-chat/rooms/:roomId/agents",
        "operationId": "createAgents",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "invited": {
                    "type": "boolean"
                  },
                  "name": {
                    "type": "string"
                  },
                  "profile": {
                    "type": "string"
                  }
                },
                "required": [
                  "profile"
                ]
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Get agents",
        "description": "GET /api/DiTing/group-chat/rooms/:roomId/agents",
        "operationId": "getAgents",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/DiTing/group-chat/rooms/{roomId}/agents/{agentId}": {
      "delete": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Delete :agentId",
        "description": "DELETE /api/DiTing/group-chat/rooms/:roomId/agents/:agentId",
        "operationId": "deleteAgents",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/DiTing/group-chat/rooms/{roomId}/clear-context": {
      "post": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Create clear-context",
        "description": "POST /api/DiTing/group-chat/rooms/:roomId/clear-context",
        "operationId": "createClear-context",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/DiTing/group-chat/rooms/{roomId}/clone": {
      "post": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Create clone",
        "description": "POST /api/DiTing/group-chat/rooms/:roomId/clone",
        "operationId": "createClone",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inviteCode": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/DiTing/group-chat/rooms/{roomId}/compress": {
      "post": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Create compress",
        "description": "POST /api/DiTing/group-chat/rooms/:roomId/compress",
        "operationId": "createCompress",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/DiTing/group-chat/rooms/{roomId}/config": {
      "put": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Update config",
        "description": "PUT /api/DiTing/group-chat/rooms/:roomId/config",
        "operationId": "updateConfig",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxHistoryTokens": {
                    "type": "number"
                  },
                  "tailMessageCount": {
                    "type": "number"
                  },
                  "triggerTokens": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/DiTing/group-chat/rooms/{roomId}/invite-code": {
      "put": {
        "tags": [
          "Group Chat"
        ],
        "summary": "Update invite-code",
        "description": "PUT /api/DiTing/group-chat/rooms/:roomId/invite-code",
        "operationId": "updateInvite-code",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inviteCode": {
                    "type": "string"
                  }
                },
                "required": [
                  "inviteCode"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/DiTing/performance/runtime": {
      "get": {
        "tags": [
          "Performance"
        ],
        "summary": "Get runtime",
        "description": "GET /api/DiTing/performance/runtime",
        "operationId": "runtime",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad Request - Invalid parameters",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request"
                }
              }
            }
          }
        }
      },
...
    }
  }
}